Move `binarySearch` to `utils`
